🚀 chore(setup.py): add package data for web3_manager and uniswap_v3 This commit adds package data for web3_manager and uniswap_v3 packages. The assets directory is included in the package data for both packages. This is necessary for the packages to function correctly.
🔨 refactor(TokenManagement.py): add type hints and docstrings to TokenManager class methods
This commit adds type hints and docstrings to the TokenManager class methods. The docstrings provide additional information on the methods' functionality, including the formulas used to calculate the values. The type hints improve the readability of the code and make it easier to understand the expected input and output of each method.

🔥 chore(TokenManagement.py): remove commented out code
✨ feat(TokenManagement.py): add get_swap_amount method
The commented out code was removed as it was not being used and was just cluttering the file. The get_swap_amount method was added to calculate the amount of token to swap based on the amount of token X provided. This method is useful for creating a pool with a bit more than half of the other token.

🚀 feat(web_manager.py): Refactor webManager class to Web3Manager class
This commit refactors the webManager class to Web3Manager class to improve naming conventions. The Web3Manager class now has an updated constructor that takes in more parameters and initializes the Uniswap object. The open_position method now takes in a capitalPercentage parameter and returns a TxReceipt object. The close_position method now takes in a tokenId parameter and returns a TxReceipt object. The update_position method now updates the position by closing the current position if the tick is outside the range, swapping tokens, and opening a new position. The get_current_price method now returns the current price of the pool. The swap_amounts method now swaps the tokens in the wallet for the token with the least amount.