Feature token management (#6)
* ✨ feat(TokenManagement.py): add functions for Uniswap V3 token management
This commit adds functions for Uniswap V3 token management, including conversion between price and tick, calculation of prices and ticks for a range with defined percentages, calculation of liquidity, and calculation of swap amounts when opening a new LP after exiting a previous range. The functions added are `priceToTick()`, `adjustedPricesToTicks()`, `tickToPrice()`, `adjustedPricesFromTick()`, `getRanges()`, `getLiquidity()`, `getTokenAmount()`, and `getSwapAmount()`. These functions are used to manage tokens in Uniswap V3.

* 🖤 Format code with black

* 🚀 feat(TokenManagement.py): add getSwapAmount function
The getSwapAmount function calculates how much of asset y is needed to create a new pool, given the current price. It calculates the price ranges from the percentage and then calculates the liquidity for the top half of the range. Finally, it uses the liquidity to calculate the amount of asset y. This function improves the functionality of the TokenManagement module by providing a way to calculate the amount of asset y needed to create a new pool.

* 🖤 Format code with black

* 🔨 refactor(TokenManagement.py): fix typo in getSwapAmount function
The typo in the getSwapAmount function has been fixed. The calculation for amountY has been corrected to use the correct formula.

* 🔨 refactor(TokenManagement.py): refactor tickToPrice and priceToTick functions
✨ feat(TokenManagement.py): add getMinX and getSwapAmount functions
The tickToPrice and priceToTick functions have been refactored to use the square root price x 96 instead of the tick. The getMinX and getSwapAmount functions have been added to calculate the minimum amount of asset x needed to create a new pool and the amount of asset y to swap for a given amount of asset x, respectively. These functions are used in the Uniswap V3 protocol.

* 🖤 Format code with black

* 🔨 refactor(TokenManagement.py): remove unused parameters and update function name
The `getSwapAmount` function was updated to only take in the `amount` parameter and return the calculated swap amount. The `token`, `price`, and `percentage` parameters were removed as they were not being used. The function name was also updated to better reflect its purpose. Additionally, a comment was added to explain the reasoning behind the calculation.

* 🖤 Format code with black

* 🚨 test(test_uniswap_math.py): add tests for TokenManager class
Added tests for the TokenManager class to ensure that the priceToTick and tickToPrice methods are working as expected. The tests use pre-defined values for ticks and prices to check if the methods return the expected values.

* 🚀 feat(TokenManagement.py): create TokenManager class and refactor functions to be methods
The TokenManagement module has been refactored to a TokenManager class. The functions have been converted to methods of the class. This improves the organization of the code and allows for better encapsulation of the data.

* 🖤 Format code with black

* 🎉 feat(web_manager.py): add webManager class to manage Uniswap positions
The webManager class is added to manage Uniswap positions. It initializes a pool with an associated wallet and a percentage. It can open and close positions at Uniswap, and if autoSwap is set to True, it will automatically swap the tokens when the position is closed, so the pool can be reopened. If autoOpen is set to True, it will automatically open a new position if there are enough funds to do so.

* Update test_uniswap_math.py

* Feature-TokenManager-v2 (#5)

* 🚀 (uniswap_v3): Upload a customized uniswap-python library

* 🔧 chore(black.yml): change trigger to only pull_request
This commit changes the trigger for the Python Black Formatter workflow to only run on pull requests. This is done to reduce the number of unnecessary workflow runs and to ensure that the code is properly formatted before merging into the main branch.

* 🚀 chore(setup.py): add package data for web3_manager and uniswap_v3
This commit adds package data for web3_manager and uniswap_v3 packages. The assets directory is included in the package data for both packages. This is necessary for the packages to function correctly.

🔨 refactor(TokenManagement.py): add type hints and docstrings to TokenManager class methods
This commit adds type hints and docstrings to the TokenManager class methods. The docstrings provide additional information on the methods' functionality, including the formulas used to calculate the values. The type hints improve the readability of the code and make it easier to understand the expected input and output of each method.

🔥 chore(TokenManagement.py): remove commented out code
✨ feat(TokenManagement.py): add get_swap_amount method
The commented out code was removed as it was not being used and was just cluttering the file. The get_swap_amount method was added to calculate the amount of token to swap based on the amount of token X provided. This method is useful for creating a pool with a bit more than half of the other token.

🚀 feat(web_manager.py): Refactor webManager class to Web3Manager class
This commit refactors the webManager class to Web3Manager class to improve naming conventions. The Web3Manager class now has an updated constructor that takes in more parameters and initializes the Uniswap object. The open_position method now takes in a capitalPercentage parameter and returns a TxReceipt object. The close_position method now takes in a tokenId parameter and returns a TxReceipt object. The update_position method now updates the position by closing the current position if the tick is outside the range, swapping tokens, and opening a new position. The get_current_price method now returns the current price of the pool. The swap_amounts method now swaps the tokens in the wallet for the token with the least amount.

* 🖤 Format code with black

---------

Co-authored-by: adradr <adradr@users.noreply.github.com>

* 🐛 fix(web_manager.py): fix currentPrice assignment
The currentPrice variable was assigned the second element of the tuple returned by the slot0() function. This was incorrect as the first element of the tuple is the current price. The code has been updated to assign the correct value to currentPrice.

* 🐛 fix(TokenManagement.py): change return type of price_to_sqrt_price_x_96 from int to float
The return type of the price_to_sqrt_price_x_96 method has been changed from int to float to ensure that the method returns a more accurate value. This change is necessary because the method is used to convert a price to a square root price, which requires a high degree of accuracy.

* 🖤 Format code with black

---------

Co-authored-by: DiodaPatr <DiodaPatr@users.noreply.github.com>
Co-authored-by: adradr <adrian.lenard@me.com>
Co-authored-by: adradr <adradr@users.noreply.github.com>