🎉 feat(uniswap_fee_and_divergence.py): Library consists of different classes to handle the lifecycle of the complete backtesting:`DataFetcher`, `HedgeManager`, `PositionManager`, `DataBacktester`.
🎉 feat(uniswap_fee_and_divergence.ipynb): add notebook for backtesting Uniswap fee and divergence by utilzing a sophisticated backtesting library `uniswap_fee_and_divergence.py`

🎉 feat(uniswap_fee_and_divergence_combinations.ipynb): add notebook for backtesting Uniswap fee and divergence combinations
This commit adds a Jupyter notebook for backtesting Uniswap fee and divergence combinations. The notebook imports the necessary libraries and defines functions for estimating the fee multiplier, running the backtest, and saving the results to a pickle file. The backtest is run in parallel using multiprocessing to speed up the process. The results are saved to a pickle file for further analysis.

🎉 feat(uniswap_fee_and_divergence_result_analysis.ipynb): add notebook for analysing the results of the `uniswap_fee_and_divergence_combinations.ipynb` that generates a combination of backtests with and without hedge for different ranges. Results are summarized in the `backtesting/README.md`.